chore(bin/node): refactor logic to allow logic to be moved to crates
