+++ bin/node/Cargo.toml
@@ -24,6 +24,8 @@ reth.workspace = true
 reth-optimism-node.workspace = true
 reth-optimism-cli.workspace = true
 reth-cli-util.workspace = true
+reth-db.workspace = true
+reth-optimism-chainspec.workspace = true
 
 # async
 futures-util.workspace = true
@@ -36,6 +38,7 @@ clap.workspace = true
 tracing.workspace = true
 url.workspace = true
 once_cell.workspace = true
+eyre.workspace = true
 
 [features]
 default = []
