+++ bin/node/README.md
@@ -1,0 +1,4 @@
+## Base Reth Node
+
+This is the main entry point for running a Base node with Reth,
+including support for flashblocks, transaction tracing, and metering.
