+++ bin/node/src/main.rs
@@ -1,177 +1,25 @@
-//! Base Reth Node binary.
-//!
-//! This is the main entry point for running a Base node with Reth,
-//! including support for flashblocks, transaction tracing, and metering.
+#![doc = include_str!("../README.md")]
+#![doc(issue_tracker_base_url = "https://github.com/base/node-reth/issues/")]
+#![cfg_attr(docsrs, feature(doc_cfg, doc_auto_cfg))]
+#![cfg_attr(not(test), warn(unused_crate_dependencies))]
 
-use std::sync::Arc;
-
-use base_reth_flashblocks_rpc::{
-    rpc::{EthApiExt, EthApiOverrideServer},
-    state::FlashblocksState,
-    subscription::FlashblocksSubscriber,
-};
-use base_reth_metering::{MeteringApiImpl, MeteringApiServer};
-use base_reth_transaction_status::{TransactionStatusApiImpl, TransactionStatusApiServer};
-use base_reth_transaction_tracing::transaction_tracing_exex;
 use clap::Parser;
-use futures_util::TryStreamExt;
-use once_cell::sync::OnceCell;
-use reth::{
-    builder::{EngineNodeLauncher, Node, NodeHandle, TreeConfig},
-    providers::providers::BlockchainProvider,
-    version::{RethCliVersionConsts, default_reth_version_metadata, try_init_version_metadata},
-};
-use reth_exex::ExExEvent;
 use reth_optimism_cli::{Cli, chainspec::OpChainSpecParser};
-use reth_optimism_node::OpNode;
-use tracing::info;
-use url::Url;
 
 pub mod cli;
-pub use cli::Args;
-
-/// The client version string for the Base Reth node.
-pub const NODE_RETH_CLIENT_VERSION: &str = concat!("base/v", env!("CARGO_PKG_VERSION"));
+pub mod node;
+pub mod version;
 
 #[global_allocator]
 static ALLOC: reth_cli_util::allocator::Allocator = reth_cli_util::allocator::new_allocator();
 
 fn main() {
-    let default_version_metadata = default_reth_version_metadata();
-    try_init_version_metadata(RethCliVersionConsts {
-        name_client: "Base Reth Node".to_string().into(),
-        cargo_pkg_version: format!(
-            "{}/{}",
-            default_version_metadata.cargo_pkg_version,
-            env!("CARGO_PKG_VERSION")
-        )
-        .into(),
-        p2p_client_version: format!(
-            "{}/{}",
-            default_version_metadata.p2p_client_version, NODE_RETH_CLIENT_VERSION
-        )
-        .into(),
-        extra_data: format!("{}/{}", default_version_metadata.extra_data, NODE_RETH_CLIENT_VERSION)
-            .into(),
-        ..default_version_metadata
-    })
-    .expect("Unable to init version metadata");
+    // Step 1: Initialize versioning so logs / telemetry report the right build info.
+    version::Version::init();
 
-    Cli::<OpChainSpecParser, Args>::parse()
-        .run(|builder, args| async move {
-            info!(message = "starting custom Base node");
+    // Step 2: Parse CLI arguments and hand execution to the Optimism node runner.
+    let cli = Cli::<OpChainSpecParser, cli::Args>::parse();
 
-            let flashblocks_enabled = args.flashblocks_enabled();
-            let transaction_tracing_enabled = args.enable_transaction_tracing;
-            let metering_enabled = args.enable_metering;
-            let op_node = OpNode::new(args.rollup_args.clone());
-
-            let fb_cell: Arc<OnceCell<Arc<FlashblocksState<_>>>> = Arc::new(OnceCell::new());
-
-            let NodeHandle { node: _node, node_exit_future } = builder
-                .with_types_and_provider::<OpNode, BlockchainProvider<_>>()
-                .with_components(op_node.components())
-                .with_add_ons(op_node.add_ons())
-                .on_component_initialized(move |_ctx| Ok(()))
-                .install_exex_if(
-                    transaction_tracing_enabled,
-                    "transaction-tracing",
-                    move |ctx| async move {
-                        Ok(transaction_tracing_exex(ctx, args.enable_transaction_tracing_logs))
-                    },
-                )
-                .install_exex_if(flashblocks_enabled, "flashblocks-canon", {
-                    let fb_cell = fb_cell.clone();
-                    move |mut ctx| async move {
-                        let fb = fb_cell
-                            .get_or_init(|| {
-                                Arc::new(FlashblocksState::new(
-                                    ctx.provider().clone(),
-                                    args.max_pending_blocks_depth,
-                                ))
-                            })
-                            .clone();
-                        Ok(async move {
-                            while let Some(note) = ctx.notifications.try_next().await? {
-                                if let Some(committed) = note.committed_chain() {
-                                    for b in committed.blocks_iter() {
-                                        fb.on_canonical_block_received(b);
-                                    }
-                                    let _ = ctx.events.send(ExExEvent::FinishedHeight(
-                                        committed.tip().num_hash(),
-                                    ));
-                                }
-                            }
-                            Ok(())
-                        })
-                    }
-                })
-                .extend_rpc_modules(move |ctx| {
-                    if metering_enabled {
-                        info!(message = "Starting Metering RPC");
-                        let metering_api = MeteringApiImpl::new(ctx.provider().clone());
-                        ctx.modules.merge_configured(metering_api.into_rpc())?;
-                    }
-
-                    let proxy_api = TransactionStatusApiImpl::new(
-                        args.rollup_args.sequencer.clone(),
-                        ctx.pool().clone(),
-                    )
-                    .expect("Failed to create transaction status proxy");
-                    ctx.modules.merge_configured(proxy_api.into_rpc())?;
-
-                    if flashblocks_enabled {
-                        info!(message = "Starting Flashblocks");
-
-                        let ws_url = Url::parse(
-                            args.websocket_url
-                                .expect("WEBSOCKET_URL must be set when Flashblocks is enabled")
-                                .as_str(),
-                        )?;
-
-                        let fb = fb_cell
-                            .get_or_init(|| {
-                                Arc::new(FlashblocksState::new(
-                                    ctx.provider().clone(),
-                                    args.max_pending_blocks_depth,
-                                ))
-                            })
-                            .clone();
-                        fb.start();
-
-                        let mut flashblocks_client = FlashblocksSubscriber::new(fb.clone(), ws_url);
-                        flashblocks_client.start();
-
-                        let api_ext = EthApiExt::new(
-                            ctx.registry.eth_api().clone(),
-                            ctx.registry.eth_handlers().filter.clone(),
-                            fb,
-                        );
-
-                        ctx.modules.replace_configured(api_ext.into_rpc())?;
-                    } else {
-                        info!(message = "flashblocks integration is disabled");
-                    }
-                    Ok(())
-                })
-                .launch_with_fn(|builder| {
-                    let engine_tree_config = TreeConfig::default()
-                        .with_persistence_threshold(builder.config().engine.persistence_threshold)
-                        .with_memory_block_buffer_target(
-                            builder.config().engine.memory_block_buffer_target,
-                        );
-
-                    let launcher = EngineNodeLauncher::new(
-                        builder.task_executor().clone(),
-                        builder.config().datadir(),
-                        engine_tree_config,
-                    );
-
-                    builder.launch_with(launcher)
-                })
-                .await?;
-
-            node_exit_future.await
-        })
-        .unwrap();
+    // Step 3: Hand the parsed CLI to the node runner so it can build and launch the Base node.
+    cli.run(|builder, args| async move { node::build_and_run_node(builder, args).await }).unwrap();
 }
